requirements.txt
blinker==1.8.2
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
contourpy
cycler==0.12.1
filelock==3.16.1
Flask==3.0.3
fonttools==4.57.0
fsspec==2025.3.0

idna==3.13
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib>=3.9
mpmath==1.3.0
networkx==3.1
numpy>=1.23,<1.27
opencv-python>=4.10
packaging==26.2
pillow==10.4.0
polars==1.8.2
psutil==7.2.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.4
scipy==1.10.1
six==1.17.0
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.13.2
ultralytics==8.4.45
ultralytics-thop==2.0.19
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
